edit README.md add task
